# docs/api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "伙伴匹配系统后端接口文档",
        "title": "Yupao Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/team/add": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "创建队伍",
                "parameters": [
                    {
                        "description": "创建参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.AddParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer",
                                            "format": "int64"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/delete": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "解散队伍",
                "parameters": [
                    {
                        "description": "解散参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.DeleteParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/get": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "根据 ID 获取队伍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "队伍 ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/team.Team"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/join": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "加入队伍",
                "parameters": [
                    {
                        "description": "加入参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.JoinParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "队伍列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "searchText",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/team.TeamUserVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/list/my/create": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "我创建的队伍",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/team.TeamUserVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/list/my/join": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "我加入的队伍",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/team.TeamUserVO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/list/page": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "队伍分页列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/team/quit": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "退出队伍",
                "parameters": [
                    {
                        "description": "退出参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.QuitParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/team/update": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "team"
                ],
                "summary": "更新队伍",
                "parameters": [
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/team.UpdateParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "team.AddParams": {
            "type": "object",
            "required": [
                "expireTime",
                "maxNum",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 512
                },
                "expireTime": {
                    "type": "string"
                },
                "maxNum": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "password": {
                    "type": "string",
                    "maxLength": 32
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "team.DeleteParams": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "team.JoinParams": {
            "type": "object",
            "required": [
                "teamId"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "teamId": {
                    "type": "integer"
                }
            }
        },
        "team.QuitParams": {
            "type": "object",
            "required": [
                "teamId"
            ],
            "properties": {
                "teamId": {
                    "type": "integer"
                }
            }
        },
        "team.Team": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "expireTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "maxNum": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "description": "队长",
                    "type": "integer"
                }
            }
        },
        "team.TeamUserVO": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "$ref": "#/definitions/user.User"
                },
                "description": {
                    "type": "string"
                },
                "expireTime": {
                    "type": "string"
                },
                "hasJoin": {
                    "type": "boolean"
                },
                "hasJoinNum": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "maxNum": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updateTime": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "team.UpdateParams": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 512
                },
                "expireTime": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 20
                },
                "password": {
                    "type": "string",
                    "maxLength": 32
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "types.Gender": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "GenderMale",
                "GenderFemale"
            ]
        },
...
        "user.User": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string"
                },
                "createTime": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "$ref": "#/definitions/types.Gender"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "planetCode": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "userAccount": {
                    "type": "string"
                },
                "userRole": {
                    "type": "integer"
                },
                "userStatus": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "Session cookie authentication. Example: session=your-session-id",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}